fix: screenTimeout uint8 overflow and add timeout debug logging
- Change screenTimeout from uint8_t to uint16_t to support 300s (5 min)
- uint8_t max 255 caused 300 to silently overflow to 44
- Add serial log on boot showing configured timeout value
- Add serial log when screen timeout triggers

+++ src/config/nvs_config.h
@@ -34,7 +34,7 @@ typedef struct {
 
     // Display settings
     uint8_t brightness;
-    uint8_t screenTimeout;
+    uint16_t screenTimeout;
     uint8_t rotation;       // Screen rotation (0-3)
     bool displayEnabled;
     bool invertColors;      // Invert display colors
